+++ src/main/java/krabban91/kodvent/kodvent/y2019/d05/Day5.java
@@ -9,7 +9,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class Day5 {
-    List<Integer> in;
+    List<Long> in;
 
     public Day5() {
         System.out.println("::: Starting Day 5 :::");
@@ -24,16 +24,16 @@ public class Day5 {
     }
 
     public long getPart1() {
-        LinkedList<Integer> input = new LinkedList<>();
-        input.add(1);
+        LinkedList<Long> input = new LinkedList<>();
+        input.add(1L);
         IntCodeComputer intCodeComputer = new IntCodeComputer(in, input);
         intCodeComputer.run();
         return intCodeComputer.lastOutput();
     }
 
     public long getPart2() {
-        LinkedList<Integer> input = new LinkedList<>();
-        input.add(5);
+        LinkedList<Long> input = new LinkedList<>();
+        input.add(5L);
         IntCodeComputer intCodeComputer = new IntCodeComputer(in, input);
         intCodeComputer.run();
         return intCodeComputer.lastOutput();
@@ -41,7 +41,7 @@ public class Day5 {
 
     public void readInput(String inputPath) {
         in = Stream.of(Input.getSingleLine(inputPath).split(","))
-                .map(Integer::parseInt)
+                .map(Long::parseLong)
                 .collect(Collectors.toList());
     }
 }
